separated sender and receiver
